Some header changes
I changed lines so that the messages that are sent between each program have some sort of header format, as described by the lineChangesAndformat.txt. The user can enter a custom IP and port. Type "123" as the IP for test mode, it connects to the server on the PC. I didnt finish some of the header formatting. I completely did not touch the chunks of code relating to file transfer (actually transferring bits of file) and I didnt get to another portion which I cant remember. User is told if their file name exists already and is invited to either change the file name or abort. In lineChangedAndformat.txt i say which lines I changed that were related to the header implementation ONLY.